+++ CHANGELOG.md
@@ -2,8 +2,25 @@
 
 All notable changes to the "arcaea-vns-syntax" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
 
-- Initial release
+## [0.0.2] - 2022-07-30
+
+### Added
+
+- File usage check for image and music
+
+## Fixed
+
+- Value check: only the first command of each kind were checked
+
+## [0.0.1] - 2022-07-29
+
+Initial release version of "arcaea-vns-syntax".
+
+### Features
+
+- Syntax highlight
+- Basic snippets
+- .vns file parsing
+- Basic value check